+++ spec/rubocop/cop/rspec/empty_example_group_spec.rb
@@ -268,8 +268,10 @@ RSpec.describe RuboCop::Cop::RSpec::EmptyExampleGroup do
   end
 
   context 'when a custom include method is specified' do
-    let(:cop_config) do
-      { 'CustomIncludeMethods' => %w[it_has_special_behavior] }
+    before do
+      all_cops_config
+        .dig('RSpec', 'Language', 'Includes', 'Example')
+        .push('it_has_special_behavior')
     end
 
     it 'ignores an empty example group with a custom include' do